Permisions updated

+++ webapp17/src/main/java/es/codeurjc/webapp17/controller/UsrController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import es.codeurjc.webapp17.entity.Usr;
 import es.codeurjc.webapp17.service.UsrService;
+import jakarta.servlet.http.HttpSession;
 
 @Controller
 public class UsrController {
@@ -41,16 +42,21 @@ public class UsrController {
      * @return a redirect to the admin page after attempting to delete the user
      */
     @PostMapping("user/{id}/delete")
-    public String deletePost(@PathVariable Long id) {
+    public String deletePost(@PathVariable Long id,HttpSession session) {
         // Get the user to delete
         Usr user = UsrService.findUsrById(id);
+
+        Usr userC = (Usr) session.getAttribute("user");
+       
+        if (userC.getRole() != Usr.Role.ADMIN) {
+               return "redirect:/no_admin"; // Redirect if not admin
+        }
 
         // Check if the user's email is 'superadmin@superadmin'
         if ("superadmin@superadmin".equals(user.getEmail())) {
             // Redirect with an error message or handle the case differently
             return "redirect:/admin"; // Prevent deletion of the superadmin
         }
-
         // Proceed with deletion if not the superadmin
         UsrService.deleteUsr(id);
         return "redirect:/admin"; // Redirect to the admin page after deleting the user
